docker magic
Exploit docker build variables.

Before, I had a nice little devenv container.

I would reuse it for "dist" builds as well, by
invoking `docker run ... ./build.sh` and then `docker cp`
to get the build out. But this has the downside of rebuilding
every time, even if source doesn't change.

Now we look to see if the BUILD build-arg is set. If it is,
we do the build as part of the container image build
(meaning it gets cached!)

But I still have a functional devenv container that doesn't
require a build of the source to get into the devenv.